agent loop refactored
